+++ src/IndexedLargeDiskData/Caching/CachedFile.cs
@@ -10,7 +10,7 @@ namespace IndexedLargeDiskData.Caching;
 public sealed class CachedFile : IDisposable
 {
     private readonly BlockCache _cache;
-    private long _length;
+    private ulong _length;
     private int _disposed;
 
     internal CachedFile(BlockCache cache, int id, SafeFileHandle handle, string path)
@@ -19,7 +19,7 @@ public sealed class CachedFile : IDisposable
         Id = id;
         Handle = handle;
         Path = path;
-        _length = RandomAccess.GetLength(handle);
+        _length = (ulong)RandomAccess.GetLength(handle);
     }
 
     /// <summary>Gets the cache-local identifier for this file.</summary>
@@ -29,24 +29,24 @@ public sealed class CachedFile : IDisposable
     public string Path { get; }
 
     /// <summary>Gets the length of the file in bytes as last observed by the cache.</summary>
-    public long Length => Interlocked.Read(ref _length);
+    public ulong Length => Interlocked.Read(ref _length);
 
     internal SafeFileHandle Handle { get; }
 
     /// <summary>Records that the file has grown, so cached blocks can be re-read at their new length.</summary>
     /// <param name="length">The new file length in bytes.</param>
-    public void SetLength(long length) => Interlocked.Exchange(ref _length, length);
+    public void SetLength(ulong length) => Interlocked.Exchange(ref _length, length);
 
     /// <summary>Drops every cached block belonging to this file.</summary>
     public void InvalidateAll() => _cache.InvalidateFile(Id);
 
     /// <summary>Drops the cached copy of a single block.</summary>
     /// <param name="blockIndex">The zero-based block index within this file.</param>
-    public void Invalidate(long blockIndex) => _cache.Invalidate(Id, blockIndex);
+    public void Invalidate(ulong blockIndex) => _cache.Invalidate(Id, blockIndex);
 
     /// <summary>Acquires a pinned lease on one block, reading it from disk if it is not cached.</summary>
     /// <param name="blockIndex">The zero-based block index within this file.</param>
-    public BlockLease Acquire(long blockIndex) => _cache.Acquire(this, blockIndex);
+    public BlockLease Acquire(ulong blockIndex) => _cache.Acquire(this, blockIndex);
 
     /// <inheritdoc />
     public void Dispose()
